+++ IB.Api.Client/Client/Model/PortfolioUpdate.cs
@@ -2,13 +2,15 @@ using System;
 using Newtonsoft.Json;
 
 namespace IB.Api.Client.Model
-{    public class PortfolioUpdate
+{
+    public class PortfolioUpdate
     {
         [JsonProperty("updatedOn")]
         public DateTime UpdatedOn;
 
         [JsonProperty("updated")]
-        public string Updated {
+        public string Updated
+        {
             get
             {
                 return $"{UpdatedOn.ToShortDateString()} {UpdatedOn.ToShortTimeString()} ";
@@ -22,16 +24,16 @@ namespace IB.Api.Client.Model
         public string Symbol { get; set; }
 
         [JsonProperty("position")]
-        public double Position { get; set; }
+        public decimal Position { get; set; }
 
         [JsonProperty("unrealizedPnl")]
-        public double UnrealizedPnl { get; set; }
-        
+        public decimal UnrealizedPnl { get; set; }
+
         [JsonProperty("contractId")]
         public int ContractId { get; set; }
 
         [JsonProperty("realizedPnl")]
-        public double RealizedPnl { get; internal set; }
+        public decimal RealizedPnl { get; internal set; }
 
         [JsonProperty("accountName")]
         public string AccountName { get; internal set; }
